When the user signs up, creates a user collection in firestore.
Check firestore db in case user already exists before creating a collection.

Removed link with popup.

Set up a skeleton reset password feature. Created a context to allow the user to call the reset password form from a button in the login container. I want to email the user to confirm a password change. I needed to create an object in the firestore db, which will hold user data, such as the user email.