Updated the like feature.
When the user likes a post, it increases the number of likes to the post and adds the post to the user's like history.

If the user clicks the like button again, it will remove their previous like and their like history will not contain that post.

Liking a post should add 1 to the total likes, switching from dislike to like should add 2.