Fixed bugs with like feature.
When the user clicks on the like or dislike button, it will add the disable class to the button until the like has been added/removed.

Clicking on the title of the post takes the user to the post page, instead of clicking anywhere on the post.

The handleLike implementation from DisplayPosts was added to the DisplayPost module too.