Refactored to make fewer calls to firestore.
Liking a post won't rerender from firestore. It will update the user's like history and the number of likes on the post, but the rerender is done locally by the handleLocalized functions.

Commented out the useCallback function while I reduced the number of times firestore was called on initial load. Because of React's useStrict mode, components would mount twice on load, which could potentially double or quadruple the calls to the db. I changed it so that when the DisplayPosts component mounts, it calls alters a count state, which calls another hook that ensures firebase is only called once.

I experimented with refactoring the updateLike functions, but that was fruitless so I changed it back.