In the midst of building an edit post component.
Need to edit the post in firebase on submission and close all relevant contexts.

Need to build an edit comment component that does functionally the same thing as the edit post component except for comments.

Going to refactor the render-from-database module. Too many redundancies in BusinessDB, GamingDB, etc. Going to make a CreatePost component that can consolidate the work.